% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.opsworkscm_operations.R
\name{associate_node}
\alias{associate_node}
\title{Associates a new node with the server}
\usage{
associate_node(
  ServerName = "string",
  NodeName = "string",
  EngineAttributes = list(
    list(
      Name = "string",
      Value = "string"
    )
  )
)
}
\arguments{
\item{ServerName}{[required] The name of the server with which to associate the node.}

\item{NodeName}{[required] The name of the node.}

\item{EngineAttributes}{[required] Engine attributes used for associating the node.

\strong{Attributes accepted in a AssociateNode request for Chef}
\itemize{
\item \code{CHEF_ORGANIZATION}: The Chef organization with which the node is associated. By default only one organization named \code{default} can exist.
\item \code{CHEF_NODE_PUBLIC_KEY}: A PEM-formatted public key. This key is required for the \code{chef-client} agent to access the Chef API.
}

\strong{Attributes accepted in a AssociateNode request for Puppet}
\itemize{
\item \code{PUPPET_NODE_CSR}: A PEM-formatted certificate-signing request (CSR) that is created by the node.
}}
}
\description{
Associates a new node with the server. For more information about how to disassociate a node, see DisassociateNode.
}
\details{
On a Chef server: This command is an alternative to \code{knife bootstrap}.

Example (Chef): \code{aws opsworks-cm associate-node --server-name MyServer --node-name MyManagedNode --engine-attributes "Name=CHEF_ORGANIZATION,Value=default" "Name=CHEF_NODE_PUBLIC_KEY,Value=public-key-pem"MyServer --node-name MyManagedNode --engine-attributes "Name=CHEF_ORGANIZATION,Value=default" "Name=CHEF_NODE_PUBLIC_KEY,Value=public-key-pem"}

On a Puppet server, this command is an alternative to the \code{puppet cert sign} command that signs a Puppet node CSR.

Example (Chef): \code{aws opsworks-cm associate-node --server-name MyServer --node-name MyManagedNode --engine-attributes "Name=PUPPET_NODE_CSR,Value=csr-pem"MyServer --node-name MyManagedNode --engine-attributes "Name=PUPPET_NODE_CSR,Value=csr-pem"}

A node can can only be associated with servers that are in a \code{HEALTHY} state. Otherwise, an \code{InvalidStateException} is thrown. A \code{ResourceNotFoundException} is thrown when the server does not exist. A \code{ValidationException} is raised when parameters of the request are not valid. The AssociateNode API call can be integrated into Auto Scaling configurations, AWS Cloudformation templates, or the user data of a server's instance.
}
\examples{

}
