% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.opsworkscm_operations.R
\name{delete_backup}
\alias{delete_backup}
\title{Deletes a backup}
\usage{
delete_backup(
  BackupId = "string"
)
}
\arguments{
\item{BackupId}{[required] The ID of the backup to delete. Run the DescribeBackups command to get a list of backup IDs. Backup IDs are in the format \code{ServerName-yyyyMMddHHmmssSSS}.}
}
\description{
Deletes a backup. You can delete both manual and automated backups. This operation is asynchronous.
}
\details{
An \code{InvalidStateException} is thrown when a backup deletion is already in progress. A \code{ResourceNotFoundException} is thrown when the backup does not exist. A \code{ValidationException} is thrown when parameters of the request are not valid.
}
\examples{

}
