% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.opsworkscm_operations.R
\name{describe_account_attributes}
\alias{describe_account_attributes}
\title{Describes your account attributes, and creates requests to increase limits before they are reached or exceeded}
\usage{
describe_account_attributes()
}
\description{
Describes your account attributes, and creates requests to increase limits before they are reached or exceeded.
}
\details{
This operation is synchronous.
}
\section{Accepted Parameters}{
\preformatted{describe_account_attributes()
}
}

