% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.opsworkscm_operations.R
\name{describe_backups}
\alias{describe_backups}
\title{Describes backups}
\usage{
describe_backups(BackupId = NULL, ServerName = NULL,
  NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{BackupId}{Describes a single backup.}

\item{ServerName}{Returns backups for the server with the specified ServerName.}

\item{NextToken}{This is not currently implemented for \code{DescribeBackups} requests.}

\item{MaxResults}{This is not currently implemented for \code{DescribeBackups} requests.}
}
\description{
Describes backups. The results are ordered by time, with newest backups first. If you do not specify a BackupId or ServerName, the command returns all backups.
}
\details{
This operation is synchronous.

A \code{ResourceNotFoundException} is thrown when the backup does not exist. A \code{ValidationException} is raised when parameters of the request are not valid.
}
\section{Accepted Parameters}{
\preformatted{describe_backups(
  BackupId = "string",
  ServerName = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

