% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.opsworkscm_operations.R
\name{describe_events}
\alias{describe_events}
\title{Describes events for a specified server}
\usage{
describe_events(
  ServerName = "string",
  NextToken = "string",
  MaxResults = 123
)
}
\arguments{
\item{ServerName}{[required] The name of the server for which you want to view events.}

\item{NextToken}{NextToken is a string that is returned in some command responses. It indicates that not all entries have been returned, and that you must run at least one more request to get remaining items. To get remaining results, call \code{DescribeEvents} again, and assign the token from the previous results as the value of the \code{nextToken} parameter. If there are no more results, the response object's \code{nextToken} parameter value is \code{null}. Setting a \code{nextToken} value that was not returned in your previous results causes an \code{InvalidNextTokenException} to occur.}

\item{MaxResults}{To receive a paginated response, use this parameter to specify the maximum number of results to be returned with a single call. If the number of available results exceeds this maximum, the response includes a \code{NextToken} value that you can assign to the \code{NextToken} request parameter to get the next set of results.}
}
\description{
Describes events for a specified server. Results are ordered by time, with newest events first.
}
\details{
This operation is synchronous.

A \code{ResourceNotFoundException} is thrown when the server does not exist. A \code{ValidationException} is raised when parameters of the request are not valid.
}
\examples{

}
