% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.opsworkscm_operations.R
\name{describe_servers}
\alias{describe_servers}
\title{Lists all configuration management servers that are identified with your account}
\usage{
describe_servers(ServerName = NULL, NextToken = NULL,
  MaxResults = NULL)
}
\arguments{
\item{ServerName}{Describes the server with the specified ServerName.}

\item{NextToken}{This is not currently implemented for \code{DescribeServers} requests.}

\item{MaxResults}{This is not currently implemented for \code{DescribeServers} requests.}
}
\description{
Lists all configuration management servers that are identified with your account. Only the stored results from Amazon DynamoDB are returned. AWS OpsWorks CM does not query other services.
}
\details{
This operation is synchronous.

A \code{ResourceNotFoundException} is thrown when the server does not exist. A \code{ValidationException} is raised when parameters of the request are not valid.
}
\section{Accepted Parameters}{
\preformatted{describe_servers(
  ServerName = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

