% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.opsworkscm_package.R
\docType{package}
\name{paws.opsworkscm-package}
\alias{paws.opsworkscm}
\alias{paws.opsworkscm-package}
\title{paws.opsworkscm: AWS OpsWorks for Chef Automate}
\description{
AWS OpsWorks for configuration management (CM) is a
service that runs and manages configuration management servers. You
can use AWS OpsWorks CM to create and manage AWS OpsWorks for Chef
Automate and AWS OpsWorks for Puppet Enterprise servers, and add or
remove nodes for the servers to manage.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
