% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.opsworkscm_operations.R
\name{update_server}
\alias{update_server}
\title{Updates settings for a server}
\usage{
update_server(
  DisableAutomatedBackup = TRUE|FALSE,
  BackupRetentionCount = 123,
  ServerName = "string",
  PreferredMaintenanceWindow = "string",
  PreferredBackupWindow = "string"
)
}
\arguments{
\item{DisableAutomatedBackup}{Setting DisableAutomatedBackup to \code{true} disables automated or scheduled backups. Automated backups are enabled by default.}

\item{BackupRetentionCount}{Sets the number of automated backups that you want to keep.}

\item{ServerName}{[required] The name of the server to update.}

\item{PreferredMaintenanceWindow}{}

\item{PreferredBackupWindow}{}
}
\description{
Updates settings for a server.
}
\details{
This operation is synchronous.
}
\examples{

}
